refactor: :recycle: move all `check_data()` internals into same file
